Convert integer to Long (#54)
* Convert integer to Long

* Fix all occurances

* Add a bit of formatting

---------

+++ app/src/main/java/com/example/myapplication/recipes/presentation/editrecipe/AddEditRecipeViewModel.kt
@@ -33,12 +33,12 @@ class AddEditRecipeViewModel @Inject constructor(
     private val _eventFlow = MutableSharedFlow<UiEvent>()
     val eventFlow = _eventFlow.asSharedFlow()
 
-    private var currentRecipeId: Int? = null
+    private var currentRecipeId: Long? = null
 
     init {
-        savedStateHandle.get<Int>("recipeId")?.let {
+        savedStateHandle.get<Long>("recipeId")?.let {
                 recipeId ->
-            if (recipeId != -1) {
+            if (recipeId != -1L) {
                 viewModelScope.launch {
                     recipesUseCases.getRecipe(recipeId)?.also {
                             recipe ->
